:recycle: Refactor: selectModal - dialog로 변경 및 닫기 접근성 개선
