+++ src/hooks/dialog/useEscDialog.ts
@@ -1,12 +1,10 @@
-import { SetStateAction, useEffect } from 'react';
+import { useEffect } from 'react';
 
-export default function useEsDialog(
-  setIsOpenDialog: React.Dispatch<SetStateAction<boolean>>,
-) {
+export default function useEscDialog(closeDialog: () => void) {
   useEffect(() => {
     const escDialog = (e: KeyboardEvent) => {
       if (e.keyCode === 27) {
-        setIsOpenDialog(false);
+        closeDialog();
       }
     };
