+++ src/pages/Detail/FeedItem/FeedItem.tsx
@@ -9,7 +9,7 @@ import useAuthContext from '../../../hooks/useAuthContext';
 import ChangeAlbumModal from '../../../components/Modal/ChangeAlbumModal/ChangeAlbumModal';
 import GetAccordionData from '../../../components/Upload/GetAccordionData';
 import DeleteFeedModal from './DeleteFeedModal';
-import Modal from '../../../components/Modal/SelectModal/SelectModal';
+import SelectModal from './SelectModal';
 import Carousel from '../../../components/Carousel/Carousel';
 import LoadingComponent from '../../../components/Loading/LoadingComponent';
 import StyledFeedItem from './StyledFeedItem';
@@ -74,7 +74,7 @@ export default function FeedItem() {
     setIsModalOpen(true);
   };
 
-  const handleCloseModal = () => {
+  const closeSelectModal = () => {
     setIsModalOpen(false);
   };
 
@@ -145,10 +145,10 @@ export default function FeedItem() {
             {time.replace(/-/gi, '.')}
           </time>
           {isModalOpen && (
-            <Modal
+            <SelectModal
               setDeleteModalOpen={setDeleteModalOpen}
               setChangeAlbumModalOpen={setChangeAlbumModalOpen}
-              onClose={handleCloseModal}
+              closeModal={closeSelectModal}
             />
           )}
           {deleteModalOpen && (
