+++ src/utils/dialog.ts
@@ -1,0 +1,10 @@
+const closeDialogOnClick = (
+  e: React.MouseEvent<HTMLDialogElement>,
+  closeDialog: () => void,
+) => {
+  if (e.target === e.currentTarget) {
+    closeDialog();
+  }
+};
+
+export { closeDialogOnClick };
